requirements.txt
streamlit
streamlit-lottie
transformers
sentence-transformers
faiss-cpu
googletrans==4.0.0-rc1
pymongo
requests
python-dotenv